% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{yadngd}
\alias{yadngd}
\title{A graphic device that does nothing}
\usage{
yadngd(welcome_message, show_message)
}
\arguments{
\item{welcome_message}{A warm message to welcome you.}

\item{whether}{to show the message.}
}
\description{
A graphic device that does nothing
}
